finish

+++ src/app/Feature/Dashboard/Components/Address/address-list/address-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { AddressDto } from 'src/app/Core/Models/Dtos/addressDto.models';
 import { AddressService } from 'src/app/Core/Services/address.service';
+import { SweetAlertService } from 'src/app/Miscelaneo/sweetalert.service';
 
 @Component({
   selector: 'app-address-list',
@@ -10,7 +11,10 @@ import { AddressService } from 'src/app/Core/Services/address.service';
 export class AddressListComponent implements OnInit {
   public address: AddressDto[] = [];
 
-  constructor(private _addressService: AddressService) {}
+  constructor(
+    private _addressService: AddressService,
+    private sweetalertService: SweetAlertService
+  ) {}
 
   ngOnInit(): void {
     this.getAddress();
@@ -23,8 +27,14 @@ export class AddressListComponent implements OnInit {
   }
 
   public deleteAddress(id: string) {
-    this._addressService.delete(id).subscribe(() => {
-      this.getAddress();
-    });
+    this.sweetalertService.opensweetalertdelete('¿Estas seguro de eliminar esta dirección?')
+      .subscribe((result) => {
+        if(result){
+          this._addressService.delete(id).subscribe(() => {
+            this.sweetalertService.opensweetalertsuccess('Dirección eliminada');
+            this.getAddress();
+          });
+        }
+      });
   }
 }
