+++ src/app/Feature/Dashboard/Components/Cost-center/cost-center-list/cost-center-list.component.ts
@@ -1,6 +1,9 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { CostCenterDto } from 'src/app/Core/Models/Dtos/cost-centerDto.models';
 import { CostCenterService } from 'src/app/Core/Services/cost-center.service';
+import { SweetAlertService } from 'src/app/Miscelaneo/sweetalert.service';
+import { CostCenterFormComponent } from '../cost-center-form/cost-center-form.component';
 
 @Component({
   selector: 'app-cost-center-list',
@@ -10,7 +13,11 @@ import { CostCenterService } from 'src/app/Core/Services/cost-center.service';
 export class CostCenterListComponent implements OnInit {
   public costcenters: CostCenterDto[] = [];
 
-  constructor(private _costCenterService: CostCenterService) {}
+  constructor(
+    private _costCenterService: CostCenterService,
+    private dialog: MatDialog,
+    private sweetAlertService: SweetAlertService
+  ) {}
 
   ngOnInit(): void {
     this.getCostCenters();
@@ -23,7 +30,32 @@ export class CostCenterListComponent implements OnInit {
   }
 
   public deleteCostCenter(id: string): void {
-    this._costCenterService.delete(id).subscribe(() => {
+    this.sweetAlertService
+      .opensweetalertdelete('¿Estas seguro de eliminar este centro de costo?')
+      .subscribe((result) => {
+        if (result) {
+          this._costCenterService.delete(id).subscribe(() => {
+            this.sweetAlertService.opensweetalertsuccess(
+              'Centro de costo eliminado'
+            );
+            this.getCostCenters();
+          });
+        }
+      });
+  }
+
+  public openDialogCreate() {
+    this.dialog.open(CostCenterFormComponent);
+
+    this.dialog.afterAllClosed.subscribe(() => {
+      this.getCostCenters();
+    });
+  }
+
+  public openDialogUpdate(costCenter: CostCenterDto) {
+    this.dialog.open(CostCenterFormComponent, { data: costCenter });
+
+    this.dialog.afterAllClosed.subscribe(() => {
       this.getCostCenters();
     });
   }
