+++ src/app/Feature/Dashboard/Components/Employee/employee-list/employee-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { EmployeeDto } from 'src/app/Core/Models/Dtos/employeeDto.models';
 import { EmployeeService } from 'src/app/Core/Services/employee.service';
+import { SweetAlertService } from 'src/app/Miscelaneo/sweetalert.service';
 import { EmployeeFormComponent } from '../employee-form/employee-form.component';
 
 @Component({
@@ -15,7 +16,8 @@ export class EmployeeListComponent implements OnInit {
 
   constructor(
     private _employeeService: EmployeeService,
-    private dialog: MatDialog
+    private dialog: MatDialog,
+    private sweetAlertService: SweetAlertService
   ) {}
 
   ngOnInit(): void {
@@ -29,23 +31,33 @@ export class EmployeeListComponent implements OnInit {
   }
 
   public deleteEmployee(id: string) {
-    this._employeeService.delete(id).subscribe(() => {
-      this.getEmployees();
-    });
+    this.sweetAlertService
+      .opensweetalertdelete('¿Estas seguro de eliminar este empleado?')
+      .subscribe((result) => {
+        if (result) {
+          this._employeeService.delete(id).subscribe(() => {
+            this.sweetAlertService.opensweetalertsuccess('Empleado eliminado');
+            this.getEmployees();
+          });
+        }
+      });
   }
 
   public editEmployee(employee: EmployeeDto) {
     this.dialog.open(EmployeeFormComponent, {
-      width: '700px',
-      height: '480px',
       data: employee,
+    });
+
+    this.dialog.afterAllClosed.subscribe(() => {
+      this.getEmployees();
     });
   }
 
   public openDialog() {
-    this.dialog.open(EmployeeFormComponent, {
-      width: '700px',
-      height: '480px',
+    this.dialog.open(EmployeeFormComponent);
+
+    this.dialog.afterAllClosed.subscribe(() => {
+      this.getEmployees();
     });
   }
 }
