+++ src/app/Feature/Dashboard/Components/User/user-list/user-list.component.ts
@@ -3,6 +3,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { UserDto } from 'src/app/Core/Models/Dtos/userDto.models';
 import { UserService } from 'src/app/Core/Services/user.service';
 import { SignUpComponent } from 'src/app/Feature/Auth/sign-up/sign-up.component';
+import { SweetAlertService } from 'src/app/Miscelaneo/sweetalert.service';
 
 @Component({
   selector: 'app-user-list',
@@ -11,7 +12,11 @@ import { SignUpComponent } from 'src/app/Feature/Auth/sign-up/sign-up.component'
 })
 export class UserListComponent implements OnInit {
   public users: UserDto[] = [];
-  constructor(private userService: UserService, private dialog: MatDialog) {}
+  constructor(
+    private userService: UserService,
+    private dialog: MatDialog,
+    private sweetalertService: SweetAlertService
+  ) {}
 
   ngOnInit(): void {
     this.getUsers();
@@ -24,9 +29,16 @@ export class UserListComponent implements OnInit {
   }
 
   public deleteUser(id: string): void {
-    this.userService.delete(id).subscribe(() => {
-      this.getUsers();
-    });
+    this.sweetalertService
+      .opensweetalertdelete('¿Estas seguro de eliminar este usuario?')
+      .subscribe((result) => {
+        if (result) {
+          this.userService.delete(id).subscribe(() => {
+            this.sweetalertService.opensweetalertsuccess('Usuario eliminado');
+            this.getUsers();
+          });
+        }
+      });
   }
 
   openDialog() {
